refactor: remove unused code (#55)
